fix models, move User.php to app\Models
