+++ app/Models/AdminUser.php
@@ -2,10 +2,14 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Model;
+use App\Models\AdminUser;
 
-class AdminUser extends Model
+use Illuminate\Notifications\Notifiable;
+use Illuminate\Foundation\Auth\User as Authenticatable;
+
+class AdminUser extends Authenticatable
 {
+    use Notifiable;
    
     protected $table = 'admin_users';
     /**
@@ -17,6 +21,7 @@ class AdminUser extends Model
         'name', 'email', 'password','sex', 'birthday'
     ];
 
+    public $timestamps = true;
     /**
      * The attributes that should be hidden for arrays.
      *
