+++ app/Models/Category.php
@@ -8,7 +8,9 @@ class Category extends Model
 {
     
     protected $table = 'categories';
-    protected $fillable = 'name';
+    protected $fillable = [
+        'name'
+    ];
     public $timestamps = true;
 
     /**
