+++ app/Models/SummaryImage.php
@@ -7,6 +7,8 @@ use Illuminate\Database\Eloquent\Model;
 class SummaryImage extends Model
 {
     protected $table = 'summary_images';
-    protected $fillable = 'image';
+    protected $fillable = [
+      'image'
+    ];
     public $timestamps = true;
 }
