+++ app/Models/SummaryText.php
@@ -7,6 +7,8 @@ use Illuminate\Database\Eloquent\Model;
 class SummaryText extends Model
 {
     protected $table = 'summary_text';
-    protected $fillable = 'content';
+    protected $fillable = [
+      'content'
+    ];
     public $timestamps = true;
 }
